UIEH-1062 Rename file with a constant
