+++ src/constants/index.js
@@ -24,7 +24,7 @@ export { default as platformTypes } from './platformTypes';
 export { default as usageConsolidation } from './usageConsolidationReduxStateShape';
 export { default as ucCredentialsReduxStateShape } from './ucCredentialsReduxStateShape';
 export { default as sortOrders } from './sortOrders';
-export { default as COVERAGE_STATEMENT_VALUE_MAX_LENGTH } from './coverageStatementFields';
+export { default as COVERAGE_STATEMENT_VALUE_MAX_LENGTH } from './COVERAGE_STATEMENT_VALUE_MAX_LENGTH';
 export * from './filterConfigs';
 export * from './costPerUse';
 export * from './customLabels';
